[Common] - Fix - Project + workspace + plt header

+++ RR-Backend/src/main/java/com/scor/rr/domain/TargetBuild/PLTDetails/PLTDetailSummary.java
@@ -19,7 +19,7 @@ public class PLTDetailSummary {
 
     @Id
     @Column(name = "pltId")
-    private Integer pltId;
+    private Long pltId;
 
     @Column(name = "pltName")
     private String pltName;
