+++ RR-Backend/src/main/java/com/scor/rr/domain/TargetBuild/PLTManagerView.java
@@ -17,7 +17,7 @@ public class PLTManagerView {
 
     @Id
     @Column(name = "pltId")
-    private Integer pltId;
+    private Long pltId;
 
     @Column(name = "pltName")
     private String pltName;
